Some Renaming and cleanup
